Move option_iter to the legion-prefab crate
